Rename amqp_timer.[ch] to amqp_time.[ch].
Rename the files to better reflect the functions that they contain.
