+++ librabbitmq/CMakeLists.txt
@@ -129,7 +129,7 @@ set(RABBITMQ_SOURCES
     ${AMQP_FRAMING_C_PATH}
     amqp_api.c amqp.h amqp_connection.c amqp_mem.c amqp_private.h amqp_socket.c
     amqp_table.c amqp_url.c amqp_socket.h amqp_tcp_socket.c amqp_tcp_socket.h
-    amqp_timer.c amqp_timer.h
+    amqp_time.c amqp_time.h
     amqp_consumer.c
     ${AMQP_SSL_SRCS}
 )
